splitted 'Create Asset' steps into components
